+++ src/reagent/dom.cljs
@@ -57,8 +57,6 @@
   [this]
   (react-dom/findDOMNode this))
 
-(set! input/find-dom-node dom-node)
-
 (defn force-update-all
   "Force re-rendering of all mounted Reagent components. This is
   probably only useful in a development environment, when you want to