+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 **[compare](https://github.com/reagent-project/reagent/compare/v1.0.0-alpha1...master)**
 
-- Renamed `:function-components?` option to `:function-components` ([#496](https://github.com/reagent-project/reagent/issues/496))
+- Renamed `:functional-components?` option to `:function-components` ([#496](https://github.com/reagent-project/reagent/issues/496))
 - Added `:f>` shortcut to create Function component from ClojureScript
 function.
 - Added `:r>` (raw) shortcut to use React components, without